Create Newton Bisection
Successively use either the Bisection method or Newton method to find a root of a function within a certain tolerance. Prints results into a table, where a is initial point on interval and b is endpoint. P is the current point and f(p) is the value of the current point.